Updated default config to use SQLite
Since having to fix the issues with the MySQL gem is causing so many
issues, the default db config that is generated by the setup script has
been updated to use SQLite instead. The Gemfile has also been changed to
match. This means that this part of the test no longer requires the
candidate to do anything.